+++ HISTORY.md
@@ -2,8 +2,10 @@
 Release Date: 2019-10-14
 
 * Issue 260: Improved social security number (SSN) generation that should pass basic validation by avoiding invalid SSN ranges. Note: Deterministic SSNs generated with `Person.Ssn()` will change.
-* PR 261: Added `Internet.IpAddress()`, `Internet.Ipv6Address()`, `Internet.IpEndPoint()`, and `Internet.Ipv6EndPoint()`.
 * Issue 252: `Internet.Ip()` now avoids generating IP addresses with a leading zero. For example, Bogus will not generate an IP address with a leading zero like 0.1.2.3. Note: Deterministic IPs generated with `Internet.Ip()` will change.
+* PR 261: Added `Internet.IpAddress()`, `Internet.Ipv6Address()`, `Internet.IpEndPoint()`, and `Internet.Ipv6EndPoint()`.
+* Issue 258: Add `Internet.UrlRootedPath()` to generate random `/foo/bar` paths.
+* Added `Internet.UrlWithPath(fileExt:".txt")` fileExt extension parameter to generate URLs with a specific file extension.
 
 ## v28.3.2
 Release Date: 2019-10-04